데이터 버전 확인 시 문제가 발생할 수 있는 가능성 제거

+++ GL2D/DataUtil.cpp
@@ -72,7 +72,7 @@ void DataUtil::Init() {
 }
 
 void DataUtil::CheckDataVersion() {
-	if (VERSION <= LoadData(DataVersionStr))
+	if (VERSION <= (int)LoadData(DataVersionStr))
 		return;
 
 	for (auto& D : DataList) {
